Switch to a fibre-per-peer model in the server
The server now spins off a low-priority fibre for processing handshakes.
When that fibre successfully completes a handshake with a peer, it spins
off a fibre for handling incoming packets from that peer.  The
datagram-handling fibre only calls recvfrom and puts the received
datagram into the correct queue (hshakeq or peer->recvq).

+++ include/fibre.h
@@ -31,7 +31,7 @@ extern void fibre_finish(void);
 extern int fibre_yield(void);
 extern void fibre_return(void);
 extern int fibre_current(void);
-extern void fibre_go(int prio, void (*)(int, void *), int, void *);
+extern void fibre_go(int prio, void (*)(long, void *), long, void *);
 extern int fibre_awaitfd(int fd, int events);
 extern void fibre_close(int fd);
 extern int fibre_sleep(const struct timespec *);
